refactor(codex-usage): split querying and formatting
